Update index.html
Hey Laura,

Great job on the website! I felt like the code was well written and it is one of the more visually appealing sites that I have seen you make. That being said, I still think that there is room for improvement when it comes to the look and feel of the site that could help modernize it. Keep your eye out for websites that you think are particularly well designed and have a modern feel to them. Then, use it as inspiration for future websites that you develop.

I made a few different changes in this file and cleaned up the spacing. I suspect that spacing differences could just be artifacts of using an IDE. However, I switched the div that was being used to generate a line to an hr tag. Great work overall, I'm glad to see that you carried over lots of the topics we discussed last class.

Seth